adapted to moved AbstractWorkflowBasedRunConfiguration
