+++ .gitignore
@@ -11,6 +11,7 @@
 *.war
 *.zip
 hs_err_pid*
+.polyglot*
 
 # Xtext
 *._trace
