+++ bundles/de.uka.ipd.sdq.dsexplore.analysis.lqn/src/de/uka/ipd/sdq/dsexplore/analysis/lqn/AbstractLQNAnalysis.java
@@ -31,7 +31,7 @@ import de.uka.ipd.sdq.dsexplore.analysis.PCMPhenotype;
 import de.uka.ipd.sdq.dsexplore.launch.DSEWorkflowConfiguration;
 import de.uka.ipd.sdq.workflow.jobs.JobFailedException;
 import de.uka.ipd.sdq.workflow.jobs.UserCanceledException;
-import de.uka.ipd.sdq.workflow.launchconfig.AbstractWorkflowConfigurationBuilder;
+import de.uka.ipd.sdq.workflow.launchconfig.core.AbstractWorkflowConfigurationBuilder;
 import de.uka.ipd.sdq.workflow.mdsd.blackboard.MDSDBlackboard;
 
 public abstract class AbstractLQNAnalysis extends AbstractAnalysis implements IAnalysis {
